TEST: iovec unsigned vs unsigned long compiler test
Test recommended from list...

+++ include/linux/uio.h
@@ -267,7 +267,7 @@ size_t hash_and_copy_to_iter(const void *addr, size_t bytes, void *hashp,
 		struct iov_iter *i);
 
 struct iovec *iovec_from_user(const struct iovec __user *uvector,
-		unsigned long nr_segs, unsigned long fast_segs,
+		unsigned nr_segs, unsigned fast_segs,
 		struct iovec *fast_iov, bool compat);
 ssize_t import_iovec(int type, const struct iovec __user *uvec,
 		 unsigned nr_segs, unsigned fast_segs, struct iovec **iovp,
